+++ Krowi_AchievementFilter/Localization/GlobalStrings.lua
@@ -27,12 +27,12 @@ function globalStrings.Load(L)
     L["Categories"] = CATEGORIES;
     L["Help"] = GAMEMENU_HELP;
     L["Options"] = GAMEOPTIONS_MENU;
-    L["General"] = GENERAL;
-    L["Key Binding"] = KEY_BINDING;
     L["Missing"] = ADDON_MISSING;
 
     -- Options
+    L["General"] = GENERAL;
     L["Info"] = INFO;
     L["Version"] = GAME_VERSION_LABEL;
     L["Icon"] = EMBLEM_SYMBOL;
+    L["Key Binding"] = KEY_BINDING;
 end